requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
yfinance==0.2.36
pytz==2024.1
playwright==1.40.0
playwright-stealth==1.0.6
python-dotenv==1.0.0